Pushing Monday's 4th March changes
Changed a lot of the files, we now perform the quadratic energy calibration and all files are using the 24hr data collected, which represents a big improvement over the previous 2hr data collected, and used, before. Not much left to do now!

Also added the appendix, which will boost the file size massively :/